Redid organization of bed_dr and ben_drs functions
I discovered tha the ben_ and tor_ functions are likely consuming the vast majority of the runtime. Some of these functions contain in excess of 100 matrix-matrix multiplications. To better assess the issue at hand, I reorganized the code in the two ben_ functions. I Plan on attempting to improve the ben_ and tor_ functions once I have finished updating the rotation function (I will continue updating the rotation ones for sake of consistency).